pyproject.toml
[project]
name = "trade-pilot"
version = "0.1.0"
description = "AI-Powered Trading Agent for Hyperliquid"
readme = "README.md"
authors = [
    { name = "aihes", email = "aihehe123@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.5.0",
    "langchain>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
trade-pilot = "trade_pilot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
